Finished making the RectangleGeometryTool OnToolGUI method. Note still need to do the OnDrawHandles and DrawRectangleHandle helper method.
